requirements.txt
itsdangerous>=0.24
flask>=0.10.1
sqlalchemy>=1.0.12
flask-sqlalchemy>=2.0
scrapyd>=1.1.0
lxml>=3.5.0
alembic>=0.8.3
Flask-Alembic>=1.2.1
flask-login>=0.3.2
flask-openid>=1.2.5
xlrd>=0.9.4

